Add minor changes to address code review
These changes are for https://github.com/wordpress-mobile/WordPress-iOS/pull/16007#pullrequestreview-603373314
